+++ wechatpy/__init__.py
@@ -4,10 +4,13 @@ from wechatpy.parser import parse_message  # NOQA
 from wechatpy.replies import create_reply  # NOQA
 from wechatpy.client import WeChatClient  # NOQA
 from wechatpy.exceptions import WeChatException  # NOQA
+from wechatpy.exceptions import WeChatClientException  # NOQA
 from wechatpy.oauth import WeChatOAuth  # NOQA
+from wechatpy.exceptions import WeChatOAuthException  # NOQA
 from wechatpy.pay import WeChatPay  # NOQA
+from wechatpy.exceptions import WeChatPayException  # NOQA
 from wechatpy.component import WeChatComponent  # NOQA
 
 
-__version__ = '1.0.0'
+__version__ = '1.0.4'
 __author__ = 'messense'