+++ wechatpy/client/api/__init__.py
@@ -19,3 +19,4 @@ from wechatpy.client.api.shakearound import WeChatShakeAround  # NOQA
 from wechatpy.client.api.device import WeChatDevice  # NOQA
 from wechatpy.client.api.template import WeChatTemplate  # NOQA
 from wechatpy.client.api.poi import WeChatPoi  # NOQA
+from wechatpy.client.api.wifi import WeChatWiFi  # NOQA